+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 *   (bug) Add empty extra field for slices.
 *   (improvement) Added `ImageValue` and `VideoValue`.
+*   (improvement) Implement `EmbedField` (it only supports videos for now).
 
 
 4.0.0-beta.2